+++ src/App.jsx
@@ -1,9 +1,25 @@
 import "./index.css";
+import {
+  createBrowserRouter,
+  RouterProvider,
+} from "react-router-dom";
 import Home from "./pages/Home";
-
+import Doc from "./pages/Doc";
+const router = createBrowserRouter(
+  [
+    {
+      path: "/",
+      element: <Home />,
+    },
+    {
+      path: '/doc',
+      element: <Doc />
+    }
+  ]
+);
 function App() {
   return (
-    <Home />
+    <RouterProvider router={router} />
   );
 }
